Use local variable for appending -proposed
Previously the loop across packages would use the dist package found
to determine if it was found in a -proposed pocket. If it was found,
it would simply append -proposed to the openstack release name. This
should be done in a local variable so that -proposed doesn't get
appended to subsequent packages.

Fixes #4

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>