Add cache support via the requests_cache package
This commit adds support for caching remote data using the
requests_cache package. In order to do this, the urllib library
that was being used has been switched over to use the requests
library which makes it much easier to fetch remote content.

Cache data will use the sqlite backend and store the cache data
in the ~/.cmadison directory. The sqlite database containing the
cache is cmadison.sqlite in the cache directory.

Caching is enabled by default as it is assumed that most users
would prefer to use cached data where possible. To allow users
to manage their cache and experience with the cache, two new options
are added:

  --no-cache, when specified, this flag bypasses the use of the
              local cache and fetches the data from remote URLs

  --clear-cache, when specified, this flag will remove the local
                 cache directory and all its contents

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>