Delete cache file rather than CACHE_DIR
The CACHE_DIR for snaps points to the SNAP_USER_DIR which cannot
be removed. This causes the --clear-cache flag to fail because the
code tries to simply remove the CACHE_DIR directory.

This change makes it where the sqlite database containing the
cache data is simply removed rather than the CACHE_DIR.

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>